// Engine/Source/Programs/Horde/HordeDashboard/src/backend/Api.ts
export enum JobStepState {

	// Waiting for dependencies of this step to complete (or paused)
	Waiting = "Waiting",

	// Ready to run, but has not been scheduled yet
	Ready = "Ready",

	// Dependencies of this step failed, so it cannot be executed
	Skipped = "Skipped",

	// There is an active instance of this step running
	Running = "Running",

	// This step has been run
	Completed = "Completed",

	// This step has been aborted
	Aborted = "Aborted"

}

export enum DeviceStatus {

	// Device encountered an error
	Error = "Error",

	// Device is operating normally
	Normal = "Normal"
}

// Outcome of a jobstep run
export enum JobStepOutcome {

	// Finished with errors
	Failure = "Failure",

	// Finished with warnings
	Warnings = "Warnings",

	// Finished Succesfully
	Success = "Success",

	// Unspecified, skipped, aborted, etc
	Unspecified = "Unspecified"
}
// ...
export enum EventSeverity {

	Unspecified = "Unspecified",
	Information = "Information",
	Warning = "Warning",
	Error = "Error"
}
// ...
export type EventData = GetLogEventResponse
// ...
export type LogLine = {
	time: string;
	level: LogLevel;
	message: string;
	id?: number;
	format?: string;
	properties?: Record<string, string | Record<string, string | number>>;
}
// ...
export type BatchLogEventsEntry = {
	total: number;
	events: EventData[];
}
// ...
export type GetLogEventResponse = {

	/**Unique id of the log containing this event */
	logId: string;

	/**Severity of this event */
	severity: EventSeverity;

	/**Index of the first line for this event */
	lineIndex: number;

	/**Number of lines in the event */
	lineCount: number;

	/**The issue id associated with this event */
	issueId: number | null;

	/**The structured message data for this event */
	lines: LogLine[];

}
// ...
export interface GetJobStepRefResponse {
	/**The job id */
	jobId: string;

	/**The step name */
	stepName?: string;

	/**The batch containing the step */
	batchId: string;

	/**The step identifier */
	stepId: string;

	/**The change number being built */
	change: number;

	/**The step log id */
	logId?: string;

	/**The pool id */
	poolId?: string;

	/** The agent id */
	agentId?: string;

	/**State of the step. */
	state?: JobStepState;

	/**Outcome of the step, once complete. */
	outcome?: JobStepOutcome;

	jobStartTime: Date | string;

	/**Time at which the step started. */
	startTime: Date | string;

	/**Time at which the step finished. */
	finishTime?: Date | string;

	/** Issue ids affecting this step ref */
	issueIds?: number[];

	/** Step metadata */
	metadata?: string[];

}